Fix lgtm errors and warnings

+++ modules/dispatcher/dispatch.c
@@ -393,8 +393,8 @@ static inline void re_calculate_active_dsts(ds_set_p sp)
 			oldw = dst->weight;
 			dst->weight = round(max_freeswitch_weight *
 			(1 - dst->fs_sock->stats.sess /
-			     (float)dst->fs_sock->stats.max_sess) *
-			(dst->fs_sock->stats.id_cpu / (float)100));
+			     (double)dst->fs_sock->stats.max_sess) *
+			(dst->fs_sock->stats.id_cpu / (double)100));
 
 			LM_DBG("weight update for %.*s: %d -> %d (%d %d %.3f)\n",
 			       dst->uri.len, dst->uri.s, oldw, dst->weight,
